refactor: 抽取message展示组件
